+++ packages/core/src/utils/index.ts
@@ -7,3 +7,4 @@ export * from './map'
 export * from './layer'
 export * from './variantGroup'
 export * from './warn'
+export * from './handlers'
