+++ packages/preset-mini/src/utils/handlers/index.ts
@@ -1,2 +1,7 @@
-export * from './handlers'
-export * from './shorthand'
+import { createValueHandler } from '@unocss/core'
+import * as valueHandlers from './handlers'
+
+export const handler = createValueHandler(valueHandlers)
+export const h = handler
+
+export { valueHandlers }
